Fix issue#100: CRLF across 8k boundary inserts an empty line (2nd try).
This introduces buffering of a trailing, non-leading CR for each buffer
passed to pluma_document_output_write(), for later possible recombination into
a CRLF.
Since this buffering has to be flushed at EOF, a flush method has been added to
the class. It is implicitly called at document close time.