snippets plugin: implement dual-encoding "environment"
We need to keep environment-like variables in native- and utf8-encoded byte
strings: up to now, os.environ[] was used for that purpose, but this is not
compatible with Python 3, as os.environ values are unicode strings.
This commit implements a 2-entries dictionary kept in class items for passing
our variables: entry "utf8" holds the utf-8 values and "noenc" the
native-encoded values.
These entries are itself variable dictionaries with byte string values.