Taglist: Load the local translation string from lang.mo file
When using intltool, the taglist creates a big xml file containing the
local translation string, so that it has to be compressed.

It looks like this:
```
<TagGroup sort="true" name="XHTML 1.0 - Tags">
  <Tag name="Abbreviated form">
    ...
  </Tag>
  <Tag name="...">
    ...
  </Tag>
  ...
</TagGroup>
<TagGroup xml:lang="af" sort="true" name="XHTML 1.0 - Tags">
  <Tag name="Afgekorte vorm" xml:lang="af">
    ...
  </Tag>
  <Tag name="..." xml:lang="af">
    ...
  </Tag>
  ...
</TagGroup>
<TagGroup xml:lang="..." sort="true" name="XHTML 1.0 - Tags">
</TagGroup>
```

Obviously, it wastes space and download bandwidth.

When switch from intltool to gettext, it does not generate a similar
huge xml file. It only get the translate string into pot file from the
xml file, and the original xml file has not changed.

This patch let taglist-plugin read the local translation string directly
from the mo file, so that it can work as before.

TODO: we need to improve and optimize the taglist plugin code and
drop some extra code.