Fix window size saving when maximized
It is expected when maximizing a window, that the unmaximized
size should be remembered, and restored again when the window
is unmaximized.

The current behaviour was saving the maximized size, in addition
to the maximized state, which is unexpected.

In addition, there is no reason to hardcode a default size, as
the schema contains a default size.